telegram_bot/requirements.txt
# Core Telegram Bot
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# WebSocket and API
websocket-client>=1.6.0
requests>=2.31.0
aiohttp>=3.8.0

# Trading and Finance
pandas>=2.0.0
numpy>=2.0.0

# Async and Scheduling
asyncio-mqtt>=0.11.0
schedule>=1.2.0

# Data Handling
json5>=0.9.0
python-dateutil>=2.8.0

# Logging and Utilities
loguru>=0.7.0
colorama>=0.4.6

# Development
pytest>=7.4.0
black>=23.0.0
